refactoring the code
